+++ spec/lib/concierge/environment_spec.rb
@@ -1,18 +1,21 @@
 require "spec_helper"
 
 RSpec.describe Concierge::Environment do
-  let(:path) { Hanami.root.join("spec", "fixtures", "environment.yml").to_s }
+  let(:global_path) { Hanami.root.join("spec", "fixtures", "environment.yml").to_s }
+  let(:app_path) { Hanami.root.join("spec", "fixtures", "app_environment.yml").to_s }
 
-  subject { described_class.new(path) }
+  subject { described_class.new(paths: [global_path, app_path]) }
 
   before do
     ENV["VARIABLE_A"] = "A"
     ENV["VARIABLE_B"] = "B"
+    ENV["VARIABLE_C"] = "C"
   end
 
   after do
     ENV.delete("VARIABLE_A")
     ENV.delete("VARIABLE_B")
+    ENV.delete("VARIABLE_C")
   end
 
   describe "#verify!" do
@@ -32,6 +35,10 @@ RSpec.describe Concierge::Environment do
       expect {
         subject.verify!
       }.to raise_error Concierge::Environment::UndefinedVariableError
+    end
+
+    it "removes duplications between multiple sources" do
+      expect(subject.send(:required_variables)).to eq ["VARIABLE_A", "VARIABLE_B", "VARIABLE_C"]
     end
   end
 end